# apispec_1.json
{
...
  "info": {
    "description": "powered by Flasgger",
    "termsOfService": "/tos",
    "title": "A swagger API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/publication/add": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "description": {
                  "example": "Description d\u00e9taill\u00e9e",
                  "type": "string"
                },
                "element": {
                  "example": "\u00c9l\u00e9ment associ\u00e9",
                  "type": "string"
                },
                "moderateur_id": {
                  "example": 1,
                  "type": "integer"
                },
                "nb_aime_negatif": {
                  "example": 0,
                  "type": "integer"
                },
                "nb_aime_positif": {
                  "example": 0,
                  "type": "integer"
                },
                "titre": {
                  "example": "Titre de la publication",
                  "type": "string"
                }
              },
              "required": [
                "titre",
                "description",
                "element",
                "autorite_id",
                "signalement_id"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Publication cr\u00e9\u00e9e avec succ\u00e8s"
          },
          "400": {
            "description": "Donn\u00e9es incompl\u00e8tes"
          },
          "500": {
            "description": "Erreur serveur interne"
          }
        },
        "summary": "Cr\u00e9e une nouvelle publication."
      }
    },
...
    "/api/publication/delete/{publication_id}": {
      "delete": {
        "parameters": [
          {
            "example": 1,
            "in": "path",
            "name": "publication_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "Publication supprim\u00e9e avec succ\u00e8s"
          },
          "404": {
            "description": "Publication non trouv\u00e9e"
          },
          "500": {
            "description": "Erreur serveur interne"
          }
        },
        "summary": "Supprime une publication de la base de donn\u00e9es."
      }
    },
    "/api/publication/update/{publication_id}": {
      "put": {
        "parameters": [
          {
            "example": 1,
            "in": "path",
            "name": "publication_id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "description": {
                  "example": "Nouvelle description",
                  "type": "string"
                },
                "element": {
                  "example": "Nouvel \u00e9l\u00e9ment",
                  "type": "string"
                },
                "nb_aime_negatif": {
                  "example": 2,
                  "type": "integer"
                },
                "nb_aime_positif": {
                  "example": 5,
                  "type": "integer"
                },
                "titre": {
                  "example": "Nouveau titre",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Publication mise \u00e0 jour avec succ\u00e8s"
          },
          "400": {
            "description": "Donn\u00e9es incompl\u00e8tes"
          },
          "404": {
            "description": "Publication non trouv\u00e9e"
          },
          "500": {
            "description": "Erreur serveur interne"
          }
        },
        "summary": "Met \u00e0 jour une publication existante."
      }
    },
    "/api/publication/{autorite_id}/autorites": {
      "get": {
        "parameters": [
          {
            "example": 1,
            "in": "path",
            "name": "autorite_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des publications de l'autorit\u00e9"
          }
        },
        "summary": "R\u00e9cup\u00e8re toutes les publications associ\u00e9es \u00e0 une autorit\u00e9 sp\u00e9cifique."
      }
    },
    "/api/publication/{publication_id}": {
      "get": {
        "parameters": [
          {
            "example": 1,
            "in": "path",
            "name": "publication_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "D\u00e9tails de la publication"
          },
          "404": {
            "description": "Publication non trouv\u00e9e"
          }
        },
        "summary": "R\u00e9cup\u00e8re une publication sp\u00e9cifique par son identifiant."
      }
    },
    "/api/publication/{signalement_id}/signalements": {
      "get": {
        "parameters": [
          {
            "example": 1,
            "in": "path",
            "name": "signalement_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des publications du signalement"
          }
        },
        "summary": "R\u00e9cup\u00e8re toutes les publications associ\u00e9es \u00e0 un signalement sp\u00e9cifique."
      }
    },
...
  },
  "swagger": "2.0"
}